Test PR for Docker (#11)
* Temporarily fixed the FXML version issue

* Upgraded Gradle and took out children tags in finished fxml

* Quick fix for Progress.md

* Added the gradle.yml

* Minor Changes

* Update gradle.yml

* Update gradle.yml

* Update gradle.yml

* Update gradle.yml

* Updated gradle.yml

* Updated gradle.yml Take 2

* Updated gradle.yml Take 3

* Added a Dockerfile

* Small format changes

* Rough draft of Dockerfile

* Quick fix in the FROM clause

* Quick fix 2

* Small command changes

* Commented out code

* Switched out the commented code

* Experiment

* Experiment 2

* Experiment 3

* Test on apt-get update

* Test on apt-get update 2

* Test on apt-get install javafx

* Created an image that builds

* Added more files in the zip and tar tasks in the build.gradle

* Deleted NeededDependencies.txt
Updated README.md
Set up Docker pushing and creating releases

* Test GitHub Releasing Take 1

* The best I can do with the auto-release plugin

* Test GitHub Releasing Take 1

+++ src/main/java/bomb/modules/ab/astrology/AstroSymbols.java
@@ -23,7 +23,7 @@ public enum AstroSymbols implements Index {
     LIBRA(6), SCORPIO(7), SAGITTARIUS(8), CAPRICORN(9),
     AQUARIUS(10), PISCES(11);
 
-    private final int index;
+    private final byte index;
 
     @Override
     public int getIdx() {
@@ -36,6 +36,6 @@ public enum AstroSymbols implements Index {
      * @param index Its index number found in the bomb manual page for Astrology
      */
     AstroSymbols(int index){
-        this.index = index;
+        this.index = (byte) index;
     }
 }
